Reconstruct brainpy.backend
